Moved models to schemas.py from app.py
